Reorganize backtest models
- Introduce models subpackage
- Add convenience re-exports
- Update import paths for consistency
