+++ nautilus_trader/backtest/models/__init__.pxd
@@ -1,0 +1,22 @@
+# -------------------------------------------------------------------------------------------------
+#  Copyright (C) 2015-2025 Nautech Systems Pty Ltd. All rights reserved.
+#  https://nautechsystems.io
+#
+#  Licensed under the GNU Lesser General Public License Version 3.0 (the "License");
+#  You may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at https://www.gnu.org/licenses/lgpl-3.0.en.html
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+# -------------------------------------------------------------------------------------------------
+
+from nautilus_trader.accounting.margin_models cimport MarginModel
+from nautilus_trader.backtest.models.fee cimport FeeModel
+from nautilus_trader.backtest.models.fee cimport FixedFeeModel
+from nautilus_trader.backtest.models.fee cimport MakerTakerFeeModel
+from nautilus_trader.backtest.models.fee cimport PerContractFeeModel
+from nautilus_trader.backtest.models.fill cimport FillModel
+from nautilus_trader.backtest.models.latency cimport LatencyModel
