+++ nautilus_trader/backtest/models/latency.pxd
@@ -15,29 +15,6 @@
 
 from libc.stdint cimport uint64_t
 
-from nautilus_trader.model.book cimport OrderBook
-from nautilus_trader.model.instruments.base cimport Instrument
-from nautilus_trader.model.objects cimport Money
-from nautilus_trader.model.objects cimport Price
-from nautilus_trader.model.objects cimport Quantity
-from nautilus_trader.model.orders.base cimport Order
-
-
-cdef class FillModel:
-    cdef readonly double prob_fill_on_limit
-    """The probability of limit orders filling on the limit price.\n\n:returns: `bool`"""
-    cdef readonly double prob_fill_on_stop
-    """The probability of stop orders filling on the stop price.\n\n:returns: `bool`"""
-    cdef readonly double prob_slippage
-    """The probability of aggressive order execution slipping.\n\n:returns: `bool`"""
-
-    cpdef bint is_limit_filled(self)
-    cpdef bint is_stop_filled(self)
-    cpdef bint is_slipped(self)
-    cpdef OrderBook get_orderbook_for_fill_simulation(self, Instrument instrument, Order order, Price best_bid, Price best_ask)
-
-    cdef bint _event_success(self, double probability)
-
 
 cdef class LatencyModel:
     cdef readonly uint64_t base_latency_nanos
@@ -48,21 +25,3 @@ cdef class LatencyModel:
     """The latency (nanoseconds) for order update messages to reach the exchange.\n\n:returns: `int`"""
     cdef readonly uint64_t cancel_latency_nanos
     """The latency (nanoseconds) for order cancel messages to reach the exchange.\n\n:returns: `int`"""
-
-
-cdef class FeeModel:
-    cpdef Money get_commission(self, Order order, Quantity fill_qty, Price fill_px, Instrument instrument)
-
-
-cdef class MakerTakerFeeModel(FeeModel):
-    pass
-
-
-cdef class FixedFeeModel(FeeModel):
-    cdef Money _commission
-    cdef Money _zero_commission
-    cdef bint _charge_commission_once
-
-
-cdef class PerContractFeeModel(FeeModel):
-    cdef Money _commission
