+++ tests/unit_tests/backtest/test_enhanced_fill_models.py
@@ -16,12 +16,11 @@
 Tests for the enhanced FillModel functionality with order book simulation.
 """
 
-
-from nautilus_trader.backtest.fill_models import BestPriceFillModel
-from nautilus_trader.backtest.fill_models import OneTickSlippageFillModel
-from nautilus_trader.backtest.fill_models import SizeAwareFillModel
-from nautilus_trader.backtest.fill_models import TwoTierFillModel
+from nautilus_trader.backtest.models import BestPriceFillModel
 from nautilus_trader.backtest.models import FillModel
+from nautilus_trader.backtest.models import OneTickSlippageFillModel
+from nautilus_trader.backtest.models import SizeAwareFillModel
+from nautilus_trader.backtest.models import TwoTierFillModel
 from nautilus_trader.common.component import TestClock
 from nautilus_trader.core.rust.model import BookType
 from nautilus_trader.model.book import OrderBook
